TGP-3441: update commodity code summary
