+++ test/viewmodels/checkAnswers/GoodsDescriptionSummarySpec.scala
@@ -20,7 +20,7 @@ import base.SpecBase
 import base.TestConstants.testRecordId
 import models.AdviceStatus.AdviceReceived
 import models.DeclarableStatus.{ImmiReady, NotReadyForUse}
-import models.ReviewReason.{Inadequate, Mismatch, Unclear}
+import models.ReviewReason.{Inadequate, Unclear}
 import models.{CheckMode, NormalMode, UserAnswers}
 import pages.goodsRecord.GoodsDescriptionPage
 import play.api.i18n.Messages
@@ -67,20 +67,6 @@ class GoodsDescriptionSummarySpec extends SpecBase {
             .onPageLoad(NormalMode, testRecordId)
             .url
 
-      }
-
-      "must render a 'Does not match' tag when reviewReason is Mismatch and declarable is NotReadyForUse" in {
-        val record = recordForTestingSummaryRows.copy(
-          reviewReason = Some(Mismatch),
-          declarable = NotReadyForUse
-        )
-
-        val row =
-          GoodsDescriptionSummary.rowUpdate(record, testRecordId, NormalMode, recordLocked = false)
-
-        row.value.content.toString must include("""<strong class="govuk-tag govuk-tag--grey">""")
-        row.value.content.toString must include(messages("goodsDescription.doesNotMatch"))
-        row.value.content.toString must include(record.goodsDescription)
       }
 
       "must render a 'Not clear' tag when reviewReason is Unclear and declarable is NotReadyForUse" in {
